refactor: decompose password auth use cases
